% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_dissimilarity}
\alias{calc_dissimilarity}
\title{Calculate graph dissimilarity}
\usage{
calc_dissimilarity(x, y = NULL)
}
\arguments{
\item{x}{a graph object or list of graph objects}

\item{y}{a graph object (if x is a graph object)}
}
\value{
the dissimilarity metric from Scheiber et al. 2017
}
\description{
Calculates dissimilarity between two graphs based on topological distance
}
