% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_path}
\alias{clean_path}
\title{Standardizes pathways stored as text}
\usage{
clean_path(path, join = "+", cause = c(">", "="), sep = c(";"))
}
\arguments{
\item{path}{a character vector describing relationships between concepts}

\item{join}{a character vector that indicates two concepts are both affected similarly}

\item{cause}{a character vector that indicates implied causal relationships}

\item{sep}{a character vector that separates different pathways in the same text}
}
\value{
a character vector the same length as the input path with standardized format
}
\description{
Given pathways stored as text, cleans and preps them to be converted to a network
}
