% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_punctuation}
\alias{clean_punctuation}
\title{Tidy up punctuation}
\usage{
clean_punctuation(x, punc, default)
}
\arguments{
\item{x}{a string}

\item{punc}{a list of punctuation marks}

\item{default}{a punctuation mark that should be standard}
}
\value{
the input string with punctuation marks replaced by the default
}
\description{
Standardizes punctuation in a string
}
\examples{
clean_punctuation("a;b/c", c("/", ";"), ";")
}
