% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_series}
\alias{create_series}
\title{Arrange graphs into a series}
\usage{
create_series(graphs, order.by, sort.order = NULL)
}
\arguments{
\item{graphs}{a list of graph objects}

\item{order.by}{a character vector of metadata by which to order and group graphs}

\item{sort.order}{a vector indicating the order in which to sort order.by}
}
\value{
a list of graph objects
}
\description{
Order graphs by metadata
}
