% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_windows}
\alias{create_windows}
\title{Create a series of graphs based on sliding windows}
\usage{
create_windows(graphs, window.size = 5, startpoint = 1)
}
\arguments{
\item{graphs}{a list of graph objects}

\item{window.size}{an integer indicating how large windows should be}

\item{startpoint}{where in the list to start the first window}
}
\value{
a list of graph objects
}
\description{
Takes the union of previous graphs within a sliding window to create a series of graphs
}
