% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fill_rows}
\alias{fill_rows}
\title{Make machine-readable ontologies}
\usage{
fill_rows(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
the same data.frame with blank rows filled in
}
\description{
Given a human-readable ontology or hierarchical dictionary, fills in blank rows
}
