% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gDist}
\alias{gDist}
\title{Graph distance}
\usage{
gDist(g, h, w1, w2, w3)
}
\arguments{
\item{g}{a graph object to compare to h}

\item{h}{a graph object to compare to g}

\item{w1}{weight of term1}

\item{w2}{weight of term2}

\item{w3}{weight of term3}
}
\value{
graph distance
}
\description{
See Schieber et al. (2017) for details
}
