% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_graph}
\alias{generate_graph}
\title{Create a network from a path description}
\usage{
generate_graph(paths)
}
\arguments{
\item{paths}{a character vector describing relationships between concepts}
}
\value{
a list of graphs the same length as input paths
}
\description{
Given an implied causal pathway described in text, creates a network
}
