% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{graph_metrics}
\alias{graph_metrics}
\title{Calculate node and edge metrics}
\usage{
graph_metrics(x, metric = "page_rank", return.df = FALSE)
}
\arguments{
\item{x}{a graph object or list of graphs}

\item{metric}{a string indicating which metric to use; see details for options}

\item{return.df}{logical; should a list of metrics be returned or a unified data frame}
}
\description{
Given a graph, calculates node and edge metrics using functions from the igraph package
}
\details{
optional metrics: page_rank, edge_betweenness, strength, alpha_centrality.
See igraph help pages for details on each function.
Note that if using edge betweenness, the inverse weight of the edges will be used.
}
