% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_graphs}
\alias{merge_graphs}
\title{Unify graphs by shared characteristics}
\usage{
merge_graphs(graphs, by)
}
\arguments{
\item{graphs}{a list of graph objects}

\item{by}{a vector indicating which graphs should be merged}
}
\value{
a graph
}
\description{
Takes the union of graphs
}
