% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{new_features}
\alias{new_features}
\title{Calculate the number and identity of new features in a graph}
\usage{
new_features(x, y = NULL, return = "counts")
}
\arguments{
\item{x}{a graph object or list of graph objects}

\item{y}{a graph object (if x is a graph object)}

\item{return}{a string indicating if 'counts' or 'features' should be returned}
}
\value{
either a list of numbers or list of feature names
}
\description{
Returns the number and/or identity of new nodes and edges between two graphs
}
