% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ontology}
\alias{ontology}
\title{Synonyms grouped by term}
\format{
A data frame 230 entries of 2 variables:
\describe{
\item{Term}{The replacement/standardized term}
\item{Synonyms}{Synonyms that should be standardized}
}
}
\usage{
ontology
}
\description{
A dataset of terms to replace synonyms with, used in the example.
}
\keyword{datasets}
