% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replace_terms}
\alias{replace_terms}
\title{Merges synonymous terms}
\usage{
replace_terms(path, terms, synonyms)
}
\arguments{
\item{path}{a character vector describing relationships between concepts}

\item{terms}{a character vector listing replacement terms}

\item{synonyms}{a character vector listing synonyms to be replaced}
}
\value{
a character vector the same length as the input path with synonyms replaced
}
\description{
Replace synonyms with standard terms in a pathway created by clean_path()
}
